fix: normalize daily quote settings and split oversized modules
