+++ lib/pages/hitokoto_settings_page_info_sections.dart
@@ -1,0 +1,88 @@
+part of 'hitokoto_settings_page.dart';
+
+extension _HitokotoSettingsPageInfoSections on _HitokotoSettingsPageState {
+  Widget _buildUsageInstructionsCard({
+    required BuildContext context,
+    required AppLocalizations l10n,
+  }) {
+    final theme = Theme.of(context);
+    final colorScheme = theme.colorScheme;
+
+    return Container(
+      width: double.infinity,
+      padding: const EdgeInsets.all(20),
+      decoration: BoxDecoration(
+        color: colorScheme.surfaceContainerHigh,
+        borderRadius: BorderRadius.circular(AppTheme.cardRadius),
+        border: Border.all(
+          color: colorScheme.outline.withAlpha(30),
+          width: 1,
+        ),
+      ),
+      child: Column(
+        crossAxisAlignment: CrossAxisAlignment.start,
+        children: [
+          Row(
+            children: [
+              Icon(
+                Icons.info_outline_rounded,
+                color: colorScheme.primary,
+                size: 20,
+              ),
+              const SizedBox(width: 8),
+              Text(
+                l10n.usageInstructions,
+                style: theme.textTheme.titleSmall?.copyWith(
+                  fontWeight: FontWeight.bold,
+                  color: colorScheme.onSurface,
+                ),
+              ),
+            ],
+          ),
+          const SizedBox(height: 12),
+          ..._buildHelpItems(context, l10n),
+        ],
+      ),
+    );
+  }
+
+  Widget _buildProviderAttributionCard({
+    required BuildContext context,
+    required AppLocalizations l10n,
+    required String providerLabel,
+  }) {
+    final theme = Theme.of(context);
+    final colorScheme = theme.colorScheme;
+
+    return Container(
+      width: double.infinity,
+      padding: const EdgeInsets.all(16),
+      decoration: BoxDecoration(
+        color: colorScheme.surfaceContainerLow,
+        borderRadius: BorderRadius.circular(AppTheme.cardRadius),
+        border: Border.all(
+          color: colorScheme.outline.withAlpha(30),
+          width: 1,
+        ),
+      ),
+      child: Row(
+        children: [
+          Icon(
+            Icons.info_outline_rounded,
+            color: colorScheme.primary,
+            size: 16,
+          ),
+          const SizedBox(width: 8),
+          Expanded(
+            child: Text(
+              l10n.dailyQuoteServiceProvider(providerLabel),
+              style: theme.textTheme.bodySmall?.copyWith(
+                color: colorScheme.onSurface.withAlpha(150),
+              ),
+            ),
+          ),
+        ],
+      ),
+    );
+  }
+}
