+++ lib/services/api_service.dart
@@ -8,6 +8,9 @@ import '../services/network_service.dart';
 import '../utils/app_logger.dart';
 import '../utils/http_response.dart';
 
+part 'api_service_daily_quote_fallback.dart';
+part 'api_service_daily_quote_remote.dart';
+
 typedef DailyQuoteHttpGet = Future<HttpResponse> Function(
   String url, {
   Map<String, String>? headers,
@@ -227,379 +230,6 @@ class ApiService {
         offlineQuoteSource: offlineQuoteSource,
       );
     }
-  }
-
-  static Future<Map<String, dynamic>?> _fetchRemoteQuote(
-    String type, {
-    required String provider,
-    required List<String> apiNinjasCategories,
-    required DailyQuoteApiKeyResolver apiKeyResolver,
-    required DailyQuoteHttpGet httpGet,
-  }) async {
-    switch (provider) {
-      case hitokotoProvider:
-        return _fetchFromHitokoto(type, httpGet: httpGet);
-      case zenQuotesProvider:
-        return _fetchFromZenQuotes(type, httpGet: httpGet);
-      case apiNinjasProvider:
-        return _fetchFromApiNinjas(
-          type,
-          apiNinjasCategories: apiNinjasCategories,
-          apiKeyResolver: apiKeyResolver,
-          httpGet: httpGet,
-        );
-      case meigenProvider:
-        return _fetchFromMeigen(type, httpGet: httpGet);
-      case koreanAdviceProvider:
-        return _fetchFromKoreanAdvice(type, httpGet: httpGet);
-      default:
-        return null;
-    }
-  }
-
-  static Future<Map<String, dynamic>?> _fetchFromHitokoto(
-    String type, {
-    required DailyQuoteHttpGet httpGet,
-  }) async {
-    String apiUrl = baseUrl;
-    if (type.contains(',')) {
-      final types = type.split(',');
-      final typeParams =
-          types.map((selectedType) => 'c=$selectedType').join('&');
-      apiUrl = '$apiUrl?$typeParams';
-    } else {
-      apiUrl = '$apiUrl?c=$type';
-    }
-
-    logDebug('一言API请求URL: $apiUrl');
-    final response = await httpGet(apiUrl, timeoutSeconds: _timeoutSeconds);
-    if (response.statusCode != 200) {
-      logDebug('一言API请求失败: ${response.statusCode}, 响应体: ${response.body}');
-      return null;
-    }
-
-    final data = json.decode(response.body);
-    if (data is! Map<String, dynamic> || !data.containsKey('hitokoto')) {
-      logDebug('一言API返回数据格式错误: $data');
-      return null;
-    }
-
-    return _normalizeQuote(
-      content: data['hitokoto']?.toString() ?? '',
-      author: data['from_who']?.toString() ?? '',
-      source: data['from']?.toString() ?? '',
-      type: data['type']?.toString() ?? _fallbackType(type),
-      provider: hitokotoProvider,
-    );
-  }
-
-  static Future<Map<String, dynamic>?> _fetchFromZenQuotes(
-    String type, {
-    required DailyQuoteHttpGet httpGet,
-  }) async {
-    const apiUrl = 'https://zenquotes.io/api/random';
-    final response = await httpGet(apiUrl, timeoutSeconds: _timeoutSeconds);
-    if (response.statusCode != 200) {
-      logDebug('ZenQuotes 请求失败: ${response.statusCode}, 响应体: ${response.body}');
-      return null;
-    }
-
-    final data = json.decode(response.body);
-    final quote = data is List && data.isNotEmpty ? data.first : data;
-    if (quote is! Map<String, dynamic>) {
-      logDebug('ZenQuotes 返回数据格式错误: $data');
-      return null;
-    }
-
-    return _normalizeQuote(
-      content: quote['q']?.toString() ?? '',
-      author: quote['a']?.toString() ?? '',
-      source: '',
-      type: _fallbackType(type),
-      provider: zenQuotesProvider,
-    );
-  }
-
-  static Future<Map<String, dynamic>?> _fetchFromApiNinjas(
-    String type, {
-    required List<String> apiNinjasCategories,
-    required DailyQuoteApiKeyResolver apiKeyResolver,
-    required DailyQuoteHttpGet httpGet,
-  }) async {
-    final apiKey = await apiKeyResolver(apiNinjasProvider);
-    if (apiKey.trim().isEmpty) {
-      logDebug('API Ninjas 未配置 API Key，跳过远程请求');
-      return null;
-    }
-
-    var apiUrl = 'https://api.api-ninjas.com/v2/randomquotes';
-    final supportedCategories = apiNinjasCategories
-        .where((category) => apiNinjasCategoryKeys.contains(category))
-        .toList();
-    if (supportedCategories.isNotEmpty) {
-      apiUrl = '$apiUrl?categories=${supportedCategories.join(',')}';
-    }
-
-    final response = await httpGet(
-      apiUrl,
-      headers: {'X-Api-Key': apiKey},
-      timeoutSeconds: _timeoutSeconds,
-    );
-    if (response.statusCode != 200) {
-      logDebug(
-          'API Ninjas 请求失败: ${response.statusCode}, 响应体: ${response.body}');
-      return null;
-    }
-
-    final data = json.decode(response.body);
-    final quote = data is List && data.isNotEmpty ? data.first : data;
-    if (quote is! Map<String, dynamic>) {
-      logDebug('API Ninjas 返回数据格式错误: $data');
-      return null;
-    }
-
-    final categories = (quote['categories'] as List<dynamic>?) ?? const [];
-    return _normalizeQuote(
-      content: quote['quote']?.toString() ?? '',
-      author: quote['author']?.toString() ?? '',
-      source: quote['work']?.toString() ?? '',
-      type: _mapApiNinjasType(categories, type),
-      provider: apiNinjasProvider,
-    );
-  }
-
-  static Future<Map<String, dynamic>?> _fetchFromMeigen(
-    String type, {
-    required DailyQuoteHttpGet httpGet,
-  }) async {
-    const apiUrl = 'https://meigen.doodlenote.net/api/json.php?c=1';
-    final response = await httpGet(apiUrl, timeoutSeconds: _timeoutSeconds);
-    if (response.statusCode != 200) {
-      logDebug('名言教えるよ 请求失败: ${response.statusCode}, 响应体: ${response.body}');
-      return null;
-    }
-
-    final data = json.decode(response.body);
-    final quote = data is List && data.isNotEmpty ? data.first : data;
-    if (quote is! Map<String, dynamic>) {
-      logDebug('名言教えるよ 返回数据格式错误: $data');
-      return null;
-    }
-
-    return _normalizeQuote(
-      content: quote['meigen']?.toString() ?? '',
-      author: quote['auther']?.toString() ?? '',
-      source: '',
-      type: _fallbackType(type),
-      provider: meigenProvider,
-    );
-  }
-
-  static Future<Map<String, dynamic>?> _fetchFromKoreanAdvice(
-    String type, {
-    required DailyQuoteHttpGet httpGet,
-  }) async {
-    const apiUrl = 'https://korean-advice-open-api.vercel.app/api/advice';
-    final response = await httpGet(apiUrl, timeoutSeconds: _timeoutSeconds);
-    if (response.statusCode != 200) {
-      logDebug(
-          'Korean Advice 请求失败: ${response.statusCode}, 响应体: ${response.body}');
-      return null;
-    }
-
-    final data = json.decode(response.body);
-    if (data is! Map<String, dynamic>) {
-      logDebug('Korean Advice 返回数据格式错误: $data');
-      return null;
-    }
-
-    return _normalizeQuote(
-      content: data['message']?.toString() ?? '',
-      author: data['author']?.toString() ?? '',
-      source: data['authorProfile']?.toString() ?? '',
-      type: _fallbackType(type),
-      provider: koreanAdviceProvider,
-    );
-  }
-
-  static String _fallbackType(String type) {
-    final normalizedType = type
-        .split(',')
-        .map((value) => value.trim())
-        .firstWhere((value) => value.isNotEmpty, orElse: () => 'a');
-    return normalizedType.isEmpty ? 'a' : normalizedType;
-  }
-
-  static String _mapApiNinjasType(
-      List<dynamic> categories, String fallbackType) {
-    for (final category in categories) {
-      switch (category.toString()) {
-        case 'philosophy':
-        case 'wisdom':
-        case 'truth':
-        case 'faith':
-          return 'k';
-        case 'art':
-        case 'writing':
-          return 'd';
-        case 'humor':
-          return 'g';
-      }
-    }
-
-    return _fallbackType(fallbackType);
-  }
-
-  static Future<String> _resolveProviderApiKey(String providerId) async {
-    return APIKeyManager().getProviderApiKey(providerId);
-  }
-
-  static Map<String, dynamic>? _normalizeQuote({
-    required String content,
-    required String author,
-    required String source,
-    required String type,
-    required String provider,
-  }) {
-    if (content.trim().isEmpty) {
-      return null;
-    }
-
-    final normalizedAuthor = author.trim();
-    final normalizedSource = source.trim();
-    final normalizedType = type.trim().isEmpty ? 'a' : type.trim();
-
-    return {
-      'content': content.trim(),
-      'source': normalizedSource,
-      'author': normalizedAuthor,
-      'type': normalizedType,
-      'from_who': normalizedAuthor,
-      'from': normalizedSource,
-      'provider': provider,
-    };
-  }
-
-  static Future<Map<String, dynamic>> _getLocalOnlyQuote(
-    AppLocalizations l10n,
-    DatabaseService? databaseService,
-    String offlineQuoteSource,
-  ) async {
-    return _getLocalQuoteOrDefault(
-      l10n,
-      databaseService,
-      offlineQuoteSource: offlineQuoteSource,
-      allowDefaultQuote: offlineQuoteSource != 'tagOnly',
-      emptyStateContent: l10n.noLocalSavedQuotes,
-      emptyStateType: 'local-empty',
-    );
-  }
-
-  // 根据 offlineQuoteSource 选择离线数据源
-  static Future<Map<String, dynamic>> _getOfflineQuote(
-    AppLocalizations l10n,
-    DatabaseService? databaseService,
-    String offlineQuoteSource, {
-    bool isOffline = false,
-  }) async {
-    return await _getLocalQuoteOrDefault(
-      l10n,
-      databaseService,
-      offlineQuoteSource: offlineQuoteSource,
-      isOffline: isOffline,
-    );
-  }
-
-  // 获取本地一言或默认一言
-  static Future<Map<String, dynamic>> _getLocalQuoteOrDefault(
-    AppLocalizations l10n,
-    DatabaseService? databaseService, {
-    String offlineQuoteSource = 'tagOnly',
-    bool isOffline = false,
-    bool allowDefaultQuote = true,
-    String? emptyStateContent,
-    String emptyStateType = 'local-empty',
-  }) async {
-    try {
-      if (databaseService != null) {
-        final localQuote = await databaseService.getLocalDailyQuote(
-          offlineQuoteSource: offlineQuoteSource,
-        );
-        if (localQuote != null) {
-          logDebug('使用本地笔记作为一言');
-          return localQuote;
-        }
-      }
-    } catch (e) {
-      logDebug('获取本地笔记失败: $e');
-    }
-
-    // 如果是离线状态且没有本地笔记，返回网络错误提示
-    if (isOffline) {
-      return {
-        'content': l10n.noNetworkConnection,
-        'source': '',
-        'author': '',
-        'type': 'offline',
-        'from_who': '',
-        'from': '',
-        'provider': 'offline',
-      };
-    }
-
-    if (!allowDefaultQuote) {
-      return {
-        'content': emptyStateContent ?? l10n.noLocalSavedQuotes,
-        'source': '',
-        'author': '',
-        'type': emptyStateType,
-        'from_who': '',
-        'from': '',
-        'provider': 'local',
-      };
-    }
-
-    // 如果不是离线状态但本地笔记获取失败，使用默认一言
-    logDebug('使用默认一言');
-    return _getDefaultQuote(l10n);
-  }
-
-  // 提供默认引言，在网络请求失败时使用
-  static Map<String, dynamic> _getDefaultQuote(AppLocalizations l10n) {
-    // 预设的引言列表
-    final quotes = [
-      {
-        'content': l10n.defaultQuote1,
-        'source': l10n.unknown,
-        'author': l10n.unknown,
-        'type': 'a',
-        'from_who': l10n.unknown,
-        'from': l10n.unknown,
-        'provider': 'default',
-      },
-      {
-        'content': l10n.defaultQuote2,
-        'source': l10n.unknown,
-        'author': l10n.unknown,
-        'type': 'a',
-        'from_who': l10n.unknown,
-        'from': l10n.unknown,
-        'provider': 'default',
-      },
-      {
-        'content': l10n.defaultQuote3,
-        'source': l10n.unknown,
-        'author': l10n.unknown,
-        'type': 'a',
-        'from_who': l10n.unknown,
-        'from': l10n.unknown,
-        'provider': 'default',
-      },
-    ];
-
-    // 随机选择一条引言
-    final random = DateTime.now().millisecondsSinceEpoch % quotes.length;
-    return quotes[random];
   }
 
   void fetchData() {
